refactor(BlackJackService): 계산결과 타입 BigDecimal로 수정

+++ src/main/java/dto/ParticipantWinning.java
@@ -1,8 +1,9 @@
 package dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record ParticipantWinning(double dealerProfit, List<PlayerProfit> playersWinning) {
+public record ParticipantWinning(BigDecimal dealerProfit, List<PlayerProfit> playersWinning) {
     public ParticipantWinning {
         playersWinning = List.copyOf(playersWinning);
     }
